refactor handleCheckout into its own controller
To make it easier to reason about as we're going to be changing this endpoint.
